+++ src/components/Header/index.tsx
@@ -6,7 +6,7 @@ import IconButton from '@mui/material/IconButton'
 import HelpIcon from '@mui/icons-material/Help'
 import SettingsIcon from '@mui/icons-material/Settings'
 import WarningRoundedIcon from '@mui/icons-material/WarningRounded'
-import { Box, Button, Chip, Menu, MenuItem, Stack, Tooltip } from '@mui/material'
+import { Box, Button, Chip, CircularProgress, Menu, MenuItem, Stack, Tooltip } from '@mui/material'
 import LoadingButton from '@mui/lab/LoadingButton'
 import { useAppStore } from '../../stores'
 import { signOut } from '../../lens/auth'
@@ -15,6 +15,7 @@ import { useAccount } from 'wagmi'
 import { ProfilePicture } from '../Shared/ProfilePicture'
 import { useQuery } from 'urql'
 import { graphql } from '../../lens/schema'
+import { SearchBar } from './SearchBar'
 
 const SettingsButton = () => {
   const showSettings = useAppStore((state) => state.showSettings)
@@ -118,7 +119,7 @@ const CurrentProfileMenu = () => {
             <CurrentProfilePicture profileId={currentProfileId} />
           </IconButton>
           :
-          <Button onClick={handleOpenMenu}>No Profile</Button>
+          <Button color='inherit' variant='outlined' onClick={handleOpenMenu}>No Profile</Button>
         }
       </Tooltip>
       <Menu
@@ -148,6 +149,7 @@ const Header = () => {
   const { address } = useAccount()
   const currentAddress = useAppStore((state) => state.currentAddress)
   const setShowBeta = useAppStore((state) => state.setShowBeta)
+  const isQuerying = useAppStore((state) => state.isQuerying)
 
   // Sign out if address changed
   useEffect(() => {
@@ -163,12 +165,16 @@ const Header = () => {
   }, [currentAddress, address])
 
   return (
-    <AppBar position='static' sx={{ zIndex: (theme) => theme.zIndex.appBar }}>
+    <AppBar position='static' color='inherit' enableColorOnDark sx={{ zIndex: (theme) => theme.zIndex.appBar }}>
       <Toolbar>
-        <Typography variant='h6' component='h1'>
-          Cultivator
-        </Typography>
-        <Chip icon={<WarningRoundedIcon />} onClick={() => setShowBeta(true)} label='Beta' color='warning' sx={{ ml: 1.5, p: 0.5 }} />
+        <Stack direction='row' alignItems='center' spacing={1.3}>
+          <Typography variant='h6' component='h1'>
+            Cultivator
+          </Typography>
+          <Chip icon={<WarningRoundedIcon />} onClick={() => setShowBeta(true)} label='Beta' color='warning' sx={{ p: 0.5 }} />
+          <SearchBar />
+          {isQuerying ? <CircularProgress size={30} /> : null}
+        </Stack>
 
         <Box sx={{ flexGrow: 1 }} />