+++ src/components/NodeDetails/Loading.tsx
@@ -1,5 +1,11 @@
+import { Box, CircularProgress } from '@mui/material'
+
 const Loading = () => {
-    return <p>Loading...</p>
+  return (
+    <Box display='flex' justifyContent='center' sx={{ p: 1 }}>
+      <CircularProgress />
+    </Box>
+  )
 }
 
 export default Loading