+++ src/constants.ts
@@ -3,7 +3,7 @@ export const APP_CHAIN_ID = process.env.REACT_APP_CHAIN_ID ? parseInt(process.en
 export const APP_CHAIN_NAME = process.env.REACT_APP_CHAIN_NAME ?? 'Polygon Mainnet'
 export const IS_MAINNET = APP_CHAIN_ID === 137
 
-export const ALCHEMY_API_KEY = process.env.REACT_APP_ALCHEMY_API_KEY
+export const ALCHEMY_API_KEY = process.env.REACT_APP_ALCHEMY_API_KEY ?? ''
 
 export const REQUEST_DELAY = 1000
 export const REQUEST_LIMIT = 50